MODULO 3/16) DATABASE MANAGER/requirements.txt
python-decouple
telethon
redis
